backend/Web_Scraping_for_Agrisense/requirements.txt
# Core dependencies for RAG data pipeline
requests==2.31.0
beautifulsoup4==4.12.3
pdfplumber==0.10.4
nltk==3.8.1
tqdm==4.66.1
lxml==5.1.0

# OCR dependencies for scanned PDFs
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.2.0

# Enhanced NLP for disease detection
spacy==3.7.2
# Run after install: python -m spacy download en_core_web_sm
